uid_t/gid_t should be formatted as unsigned long
mmkfilter_passwd_byuid()/mkfilter_group_bygid() get wrong filter string
because "%d" will return negative when uid/gid larger than 2^31, and
result to "Authentiction failure".

This also changes the other places where uid_t or gid_t values are
formatted.

+++ nslcd/cfg.c
@@ -1626,11 +1626,11 @@ static void cfg_dump(void)
   if (nslcd_cfg->uidname != NULL)
     log_log(LOG_DEBUG, "CFG: uid %s", nslcd_cfg->uidname);
   else if (nslcd_cfg->uid != NOUID)
-    log_log(LOG_DEBUG, "CFG: uid %d", (int)nslcd_cfg->uid);
+    log_log(LOG_DEBUG, "CFG: uid %lu", (unsigned long int)nslcd_cfg->uid);
   else
     log_log(LOG_DEBUG, "CFG: # uid not set");
   if (nslcd_cfg->gid != NOGID)
-    log_log(LOG_DEBUG, "CFG: gid %d", (int)nslcd_cfg->gid);
+    log_log(LOG_DEBUG, "CFG: gid %lu", (unsigned long int)nslcd_cfg->gid);
   else
     log_log(LOG_DEBUG, "CFG: # gid not set");
   log_log_config();
@@ -1783,7 +1783,7 @@ static void cfg_dump(void)
       strcpy(buffer + sizeof(buffer) - 4, "...");
     log_log(LOG_DEBUG, "CFG: nss_initgroups_ignoreusers %s", buffer);
   }
-  log_log(LOG_DEBUG, "CFG: nss_min_uid %d", (int)nslcd_cfg->nss_min_uid);
+  log_log(LOG_DEBUG, "CFG: nss_min_uid %lu", (unsigned long int)nslcd_cfg->nss_min_uid);
   log_log(LOG_DEBUG, "CFG: nss_nested_groups %s", print_boolean(nslcd_cfg->nss_nested_groups));
   log_log(LOG_DEBUG, "CFG: validnames %s", nslcd_cfg->validnames_str);
   log_log(LOG_DEBUG, "CFG: ignorecase %s", print_boolean(nslcd_cfg->ignorecase));
